fix(esp32): resolve ESP32-C6/P4 build compatibility issues
field.cpp — SECP256K1_NO_INT128 guards for 32-bit MCU paths:
  - reduce(): replace __int128 carry arithmetic with manual 64-bit
    add-with-carry using inline lambda helper (no __int128 required)
  - Step 3 overflow pass: replace u128 multiply with mul64() + carry chain

batch_verify.cpp — 32-bit narrowing conversion fixes:
  - PubkeyHash32::seed{0xcbf29ce484222325ULL} → static_cast<std::size_t>
  - PubkeyMap ctor arg: rng() → static_cast<std::size_t>(rng())

ct_point.cpp — avoid global constructor before FreeRTOS scheduler:
  - static const FieldElement B7 = from_uint64(7) runs before app_main
    on ESP32 (static-init mutex is null → crash). Replace with inline
    helper get_b7() + #define B7 get_b7() under ESP_PLATFORM guard.

tagged_hash.hpp — EspLazySHA256: constexpr-constructed lazy init:
  - 12 inline const SHA256 g_*_midstate globals run SHA-256 computations
    at startup → crash before FreeRTOS (systimer + static-init mutex).
  - ESP_PLATFORM path: EspLazySHA256 wrapper is constexpr-constructed
    (.bss, zero bytes, no ctor call). SHA-256 computed on first access
    via implicit operator const SHA256&(). Safe on single-core ESP32.

CMakeLists.txt (esp32c6_bench_hornet):
  - cpu/ → src/cpu/ path fix (directory restructure)
  - -fno-threadsafe-statics: removes mutex protection from function-local
    statics (safe on single-core, avoids ESP-IDF static-init-mutex crash)

Note: systimer-before-scheduler crash (MEPC=0x40806538) is a pre-existing
ESP-IDF v5.4 / ESP32-C6 rev v0.2 issue unrelated to this codebase.

+++ src/cpu/src/batch_verify.cpp
@@ -25,7 +25,7 @@ namespace secp256k1 {
 // Seed is randomised per batch call so adversarial pubkey inputs cannot
 // pre-compute collisions and degrade lookup to O(n) (CA-010).
 struct PubkeyHash32 {
-    std::size_t seed{0xcbf29ce484222325ULL};
+    std::size_t seed{static_cast<std::size_t>(0xcbf29ce484222325ULL)};
     explicit PubkeyHash32(std::size_t s) : seed(s) {}
     std::size_t operator()(const std::array<uint8_t, 32>& k) const noexcept {
         std::size_t h = seed;
@@ -220,7 +220,7 @@ bool schnorr_batch_verify(const SchnorrBatchEntry* entries, std::size_t n) {
     // hash collisions that degrade lookup to O(n) worst-case.
     using PubkeyMap = std::unordered_map<std::array<uint8_t, 32>, Point, PubkeyHash32>;
     static thread_local std::mt19937_64 rng{std::random_device{}()};
-    PubkeyMap pubkey_index(n, PubkeyHash32{rng()});
+    PubkeyMap pubkey_index(n, PubkeyHash32{static_cast<std::size_t>(rng())});
     pubkey_index.reserve(n);
 
     // Grow-only vector reserve (no realloc when n stays the same between calls)
